+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - add an option for multiple selection of jobs in `scheduledJobs` and then `execute` them or `delete` them
  - for `formActions` list on right click open edit or view action in a new tab
  - add `searchBuilder` for `formActions`;
+ - add `searchBuilder` for `scheduledJobs`;
 
 # v5.1.5
 ### Bugfix